feat: mission core organization, MissionUtils
