+++ Common/NPCs/SlimeGlobal.cs
@@ -1,8 +1,4 @@
-﻿using Reverie.Common.Systems;
-using Reverie.Core.Animation;
-using Reverie.Core.Missions;
-using Reverie.Core.Missions.MissionHandlers;
-using Terraria.Audio;
+﻿using Terraria.Audio;
 using Terraria.GameContent;
 using Terraria.GameContent.ItemDropRules;
