+++ Core/Missions/ObjectiveEventHandlers.cs
@@ -5,8 +5,9 @@ using Terraria.DataStructures;
 using Reverie.Core.Dialogue;
 using Reverie.Utilities.Extensions;
 using static Reverie.Core.Missions.MissionPlayer;
+using Reverie.Utilities;
 
-namespace Reverie.Core.Missions.MissionAttributes;
+namespace Reverie.Core.Missions;
 
 public class ObjectiveEventItem : GlobalItem
 {
@@ -18,18 +19,18 @@ public class ObjectiveEventItem : GlobalItem
 
     public override bool OnPickup(Item item, Player player)
     {
-        MissionHelper.TryUpdateProgressForItem(item, player);
+        MissionUtils.TryUpdateProgressForItem(item, player);
         return base.OnPickup(item, player);
     }
 
     public override void UpdateEquip(Item item, Player player)
     {
         base.UpdateEquip(item, player);
-        MissionHelper.TryUpdateProgressForItem(item, player);
+        MissionUtils.TryUpdateProgressForItem(item, player);
     }
     public override void UpdateInventory(Item item, Player player)
     {
-        MissionHelper.TryUpdateProgressForItem(item, player);
+        MissionUtils.TryUpdateProgressForItem(item, player);
         base.UpdateInventory(item, player);
     }
 }
@@ -54,7 +55,7 @@ public class ObjectiveEventNPC : GlobalNPC
     public override void EditSpawnPool(IDictionary<int, float> pool, NPCSpawnInfo spawnInfo)
     {
         var mPlayer = Main.LocalPlayer.GetModPlayer<MissionPlayer>();
-        Mission AFallingStar = mPlayer.GetMission(MissionID.AFallingStar);
+        var AFallingStar = mPlayer.GetMission(MissionID.AFallingStar);
 
         if (AFallingStar?.Progress == MissionProgress.Active)
         {
@@ -66,7 +67,7 @@ public class ObjectiveEventNPC : GlobalNPC
 
             if (!currentSet.IsCompleted && AFallingStar.CurObjectiveIndex < 2)
                 pool.Clear();
-            
+
             if (spawnInfo.PlayerInTown)
             {
                 if (AFallingStar.CurObjectiveIndex >= 3 && Main.LocalPlayer.ZoneOverworldHeight)
@@ -103,7 +104,7 @@ public class ObjectiveEventNPC : GlobalNPC
     {
         var mPlayer = Main.LocalPlayer.GetModPlayer<MissionPlayer>();
 
-        Mission AFallingStar = mPlayer.GetMission(MissionID.AFallingStar);
+        var AFallingStar = mPlayer.GetMission(MissionID.AFallingStar);
 
         if (AFallingStar?.Progress == MissionProgress.Active && player.ZoneOverworldHeight)
         {
